requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
mysqlclient==2.2.0
openai==1.3.5
python-dotenv==1.0.0